pyproject.toml
[tool.poetry]
name = "gpt-hacker"
version = "0.1.0"
description = ""
authors = ["Olivia Li <oliviali.ouyang@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
playwright = "^1.41.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
